requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
openai
faiss-cpu
python-dotenv
tiktoken
pypdf
